Renamed classes to match conventions in the COntributions file.
