+++ src/algorithms/calorimetry/CalorimeterHitDigi.cc
