+++ src/algorithms/calorimetry/CalorimeterHitDigi.h
